Fix NVFP4 4/6 GEMM availability None-safety; tidy 4/6 constants and tests
_nvfp4_availability_check crashed with AttributeError when a non-4/6 weight
quantizer had block_sizes=None (e.g. an FP8 RealQuantLinear encountered during
gemm_registry matching). Use the None-safe NVFP4QTensor._is_four_over_six helper
at both the availability check and the real-quantize scale path.

Also import FP8_E4M3_MAX in fp4_kernel_hopper from fp4_kernel (single source),
and relocate the static-quantizer 4/6 normalization-threading test from
test_mse_calibrator.py to test_nvfp4_four_over_six.py (4/6 is max-calibration
only and has no MSE-calibrator relationship).
